requirements.txt
numpy
scipy
matplotlib
pyfemm
pandas
PyYAML
OMPython
